#104 [refactor] 스케줄 서비스 리팩토링
