+++ src/main/java/hous/server/service/todo/TodoService.java
@@ -5,8 +5,6 @@ import hous.server.domain.badge.Acquire;
 import hous.server.domain.badge.BadgeInfo;
 import hous.server.domain.badge.repository.AcquireRepository;
 import hous.server.domain.badge.repository.BadgeRepository;
-import hous.server.domain.common.RedisKey;
-import hous.server.domain.room.Participate;
 import hous.server.domain.room.Room;
 import hous.server.domain.todo.Done;
 import hous.server.domain.todo.Redo;
@@ -27,12 +25,9 @@ import hous.server.service.todo.dto.request.CheckTodoRequestDto;
 import hous.server.service.todo.dto.request.TodoInfoRequestDto;
 import hous.server.service.user.UserServiceUtils;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -40,8 +35,6 @@ import java.util.List;
 @Service
 @Transactional
 public class TodoService {
-
-    private final RedisTemplate<String, Object> redisTemplate;
 
     private final UserRepository userRepository;
     private final OnboardingRepository onboardingRepository;
@@ -124,52 +117,5 @@ public class TodoService {
         doneRepository.deleteAll(todo.getDones());
         room.deleteTodo(todo);
         todoRepository.delete(todo);
-    }
-
-    @Scheduled(cron = "0  0  0  *  *  *")
-    public void scheduledDoneMyToDos() {
-        List<User> users = userRepository.findAllUsers();
-        users.forEach(user -> {
-            Onboarding onboarding = user.getOnboarding();
-            List<Participate> participates = onboarding.getParticipates();
-            if (participates.size() != 0) {
-                LocalDate yesterday = DateUtils.yesterdayLocalDate();
-                Room room = participates.get(0).getRoom();
-                if (!BadgeServiceUtils.hasBadge(badgeRepository, acquireRepository, BadgeInfo.TODO_MASTER, onboarding)) {
-                    List<Todo> todos = room.getTodos();
-                    List<Todo> allDayMyTodos = TodoServiceUtils.filterAllDaysUserTodos(todos, onboarding);
-                    List<Todo> yesterdayMyTodos = TodoServiceUtils.filterDayMyTodos(yesterday, onboarding, todos);
-                    int yesterdayDoneMyTodosCnt = (int) yesterdayMyTodos.stream()
-                            .filter(todo -> doneRepository.existsDayDoneByOnboardingAndTodo(yesterday, onboarding, todo))
-                            .count();
-                    if (allDayMyTodos.size() > 0 && yesterdayMyTodos.size() == yesterdayDoneMyTodosCnt) {
-                        String todoCompleteCountString = (String) redisTemplate.opsForValue().get(RedisKey.TODO_COMPLETE_COUNT + user.getId());
-                        if (todoCompleteCountString == null) {
-                            redisTemplate.opsForValue().set(RedisKey.TODO_COMPLETE_COUNT + user.getId(), Integer.toString(1));
-                        } else {
-                            redisTemplate.opsForValue().set(RedisKey.TODO_COMPLETE_COUNT + user.getId(), Integer.toString(Integer.parseInt(todoCompleteCountString) + 1));
-                            if (Integer.parseInt(todoCompleteCountString) == 6 && !BadgeServiceUtils.hasBadge(badgeRepository, acquireRepository, BadgeInfo.GOOD_JOB, onboarding)) {
-                                Acquire acquire = acquireRepository.save(Acquire.newInstance(onboarding, badgeRepository.findBadgeByBadgeInfo(BadgeInfo.GOOD_JOB)));
-                                onboarding.addAcquire(acquire);
-                                notificationService.sendNewBadgeNotification(user, BadgeInfo.GOOD_JOB);
-                            }
-                            if (Integer.parseInt(todoCompleteCountString) == 13 && !BadgeServiceUtils.hasBadge(badgeRepository, acquireRepository, BadgeInfo.SINCERITY_KING_HOMIE, onboarding)) {
-                                Acquire acquire = acquireRepository.save(Acquire.newInstance(onboarding, badgeRepository.findBadgeByBadgeInfo(BadgeInfo.SINCERITY_KING_HOMIE)));
-                                onboarding.addAcquire(acquire);
-                                notificationService.sendNewBadgeNotification(user, BadgeInfo.SINCERITY_KING_HOMIE);
-                            }
-                            if (Integer.parseInt(todoCompleteCountString) == 20 && !BadgeServiceUtils.hasBadge(badgeRepository, acquireRepository, BadgeInfo.TODO_MASTER, onboarding)) {
-                                Acquire acquire = acquireRepository.save(Acquire.newInstance(onboarding, badgeRepository.findBadgeByBadgeInfo(BadgeInfo.TODO_MASTER)));
-                                onboarding.addAcquire(acquire);
-                                notificationService.sendNewBadgeNotification(user, BadgeInfo.TODO_MASTER);
-                                redisTemplate.delete(RedisKey.TODO_COMPLETE_COUNT + user.getId());
-                            }
-                        }
-                    } else {
-                        redisTemplate.delete(RedisKey.TODO_COMPLETE_COUNT + user.getId());
-                    }
-                }
-            }
-        });
     }
 }
